test money

+++ src/MeneySample/Money.cs
@@ -65,8 +65,8 @@ public readonly struct Money : IComparable<Money>, IEquatable<Money>
 
   public static Money None { get; }
 
-  public static Money UnitedStatesDollar(ushort cents) => new(currency: Currency.UnitedStatesDollar, totalMinorUnits: cents);
-  public static Money Euro(ushort cents)               => new(currency: Currency.Euro              , totalMinorUnits: cents);
-  public static Money BelarusianRouble(ushort kopecks) => new(currency: Currency.BelarusianRouble  , totalMinorUnits: kopecks);
-  public static Money RussianRouble(ushort kopecks)    => new(currency: Currency.RussianRouble     , totalMinorUnits: kopecks);
+  public static Money UnitedStatesDollar(ulong cents) => new(currency: Currency.UnitedStatesDollar, totalMinorUnits: cents);
+  public static Money               Euro(ulong cents) => new(currency: Currency.Euro              , totalMinorUnits: cents);
+  public static Money BelarusianRouble(ulong kopecks) => new(currency: Currency.BelarusianRouble  , totalMinorUnits: kopecks);
+  public static Money    RussianRouble(ulong kopecks) => new(currency: Currency.RussianRouble     , totalMinorUnits: kopecks);
 }
